+++ spec/support/build_store.rb
@@ -1,8 +1,8 @@
 RSpec.configure do |c|
   c.include(Module.new do
-    def build_store(&block)
+    def build_store(partitions_count: 1, &block)
       store = Shameless::Store.new(:store) do |c|
-        c.partition_urls = ['sqlite:/']
+        c.partition_urls = Array.new(partitions_count) { 'sqlite:/' }
         c.shards_count = 4
       end